pyproject.toml
[tool.poetry]
name = "gtfs_builder"
version = "0.6.0"
description = ""
authors = ["amauryval <amauryval@gmail.com>"]
license = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.7"
geospatial_lib = "0.30.3"
starlite = "1.48.0"
uvicorn = "^0.20.0"
spatialpandas = "0.4.4"
python-dotenv = "0.15.0"


[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
